services/moneroModule/requirements.txt
Flask==1.1.2
python-monerorpc==0.6.2